Update @xmldom/xmldom
CaSS currently uses @xmldom/xmldom 0.8.3, which parses XML that is not well-formed because it contains multiple top level elements, and adds all root nodes to the `childNodes` collection of the `Document`, without reporting any error or throwing. This breaks the assumption that there is only a single root node in the tree, which led to issuance of CVE-2022-39299 as it is a potential issue for dependents.

v 0.8.4 resolves this issue.